Update with Monte Carlo Working
Please run file sim2048.py.
Monte_Carlo itself is running right now. But it does take a bit of time especially when you set the simulation number to 1000.
Also sometimes it will register a move in a direction that would produce an identical board before and after move.
The next step of this project will be counting towards the possiblity of optimizing current game, as well as fixing the repeated inavailable first move.